more patching
