+++ trulens_eval/trulens_eval/feedback/__init__.py
@@ -413,13 +413,14 @@ from trulens_eval.feedback.feedback import Feedback
 # Specific feedback functions:
 from trulens_eval.feedback.groundedness import Groundedness
 from trulens_eval.feedback.groundtruth import GroundTruthAgreement
-from trulens_eval.feedback.provider.hugs import Huggingface
+from trulens_eval.feedback.provider.cohere import Cohere
 
 # Providers of feedback functions evaluation:
+from trulens_eval.feedback.provider.hugs import Huggingface
 from trulens_eval.feedback.provider.openai import AzureOpenAI
 from trulens_eval.feedback.provider.openai import OpenAI
 
 __all__ = [
     'Feedback', 'Groundedness', 'GroundTruthAgreement', 'OpenAI', 'AzureOpenAI',
-    'Huggingface'
+    'Huggingface', 'Cohere'
 ]
