+++ trulens_eval/trulens_eval/feedback/groundedness.py
@@ -4,14 +4,15 @@ from typing import Dict, List
 import numpy as np
 from tqdm.auto import tqdm
 
-from trulens_eval.trulens_eval.feedback import _re_1_10_rating
-from trulens_eval.trulens_eval.feedback import prompts
-from trulens_eval.trulens_eval.provider import Provider
-from trulens_eval.trulens_eval.provider.hugs import Huggingface
-from trulens_eval.trulens_eval.provider.openai import AzureOpenAI
-from trulens_eval.trulens_eval.provider.openai import OpenAI
-from trulens_eval.trulens_eval.util import SerialModel
-from trulens_eval.trulens_eval.util import WithClassInfo
+
+from trulens_eval.feedback import prompts
+from trulens_eval.feedback.provider import Provider
+from trulens_eval.feedback.provider.hugs import Huggingface
+from trulens_eval.feedback.provider.openai import AzureOpenAI
+from trulens_eval.feedback.provider.openai import OpenAI
+from trulens_eval.utils.generated import re_1_10_rating
+from trulens_eval.util import SerialModel
+from trulens_eval.util import WithClassInfo
 
 logger = logging.getLogger(__name__)
 
@@ -77,7 +78,7 @@ class Groundedness(SerialModel, WithClassInfo):
             for line in reason.split('\n'):
                 if "Score" in line:
                     groundedness_scores[f"statement_{i}"
-                                       ] = _re_1_10_rating(line) / 10
+                                       ] = re_1_10_rating(line) / 10
                     i += 1
             return groundedness_scores, {"reason": reason}
         if isinstance(self.groundedness_provider, Huggingface):
