+++ trulens_eval/trulens_eval/feedback/provider/endpoint/hugs.py
@@ -4,10 +4,10 @@ from typing import Callable, Optional
 
 import requests
 
-from trulens_eval.trulens_eval.keys import _check_key
-from trulens_eval.trulens_eval.keys import get_huggingface_headers
-from trulens_eval.trulens_eval.provider.endpoint.endpoint import Endpoint
-from trulens_eval.trulens_eval.provider.endpoint.endpoint import \
+from trulens_eval.keys import _check_key
+from trulens_eval.keys import get_huggingface_headers
+from trulens_eval.feedback.provider.endpoint.endpoint import Endpoint
+from trulens_eval.feedback.provider.endpoint.endpoint import \
     EndpointCallback
 from trulens_eval.util import WithClassInfo
