+++ trulens_eval/trulens_eval/feedback/provider/hugs.py
@@ -4,13 +4,11 @@ from typing import Dict
 
 import numpy as np
 
-from trulens_eval.trulens_eval.feedback import prompts
-from trulens_eval.trulens_eval.keys import get_cohere_agent
-from trulens_eval.trulens_eval.provider import Provider
-from trulens_eval.trulens_eval.provider.endpoint.endpoint import Endpoint
-from trulens_eval.trulens_eval.provider.endpoint.endpoint import \
+from trulens_eval.feedback.provider import Provider
+from trulens_eval.feedback.provider.endpoint import Endpoint
+from trulens_eval.feedback.provider.endpoint import \
     HuggingfaceEndpoint
-from trulens_eval.trulens_eval.util import TP
+from trulens_eval.util import TP
 
 logger = logging.getLogger(__name__)
 
@@ -180,44 +178,3 @@ class Huggingface(Provider):
         for label in hf_response:
             if label['label'] == 'entailment':
                 return label['score']
-
-
-class Cohere(Provider):
-    model_engine: str = "large"
-
-    def __init__(self, model_engine='large', endpoint=None, **kwargs):
-        # NOTE(piotrm): pydantic adds endpoint to the signature of this
-        # constructor if we don't include it explicitly, even though we set it
-        # down below. Adding it as None here as a temporary hack.
-
-        kwargs['endpoint'] = Endpoint(name="cohere")
-        kwargs['model_engine'] = model_engine
-
-        super().__init__(
-            **kwargs
-        )  # need to include pydantic.BaseModel.__init__
-
-    def sentiment(
-        self,
-        text,
-    ):
-        return int(
-            Cohere().endpoint.run_me(
-                lambda: get_cohere_agent().classify(
-                    model=self.model_engine,
-                    inputs=[text],
-                    examples=prompts.COHERE_SENTIMENT_EXAMPLES
-                )[0].prediction
-            )
-        )
-
-    def not_disinformation(self, text):
-        return int(
-            Cohere().endpoint.run_me(
-                lambda: get_cohere_agent().classify(
-                    model=self.model_engine,
-                    inputs=[text],
-                    examples=prompts.COHERE_NOT_DISINFORMATION_EXAMPLES
-                )[0].prediction
-            )
-        )
